Add cast to bmpcmp.c to quiet useless compiler warning.
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10780 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/toolbin/bmpcmp.c
@@ -30,7 +30,7 @@ static void *Malloc(size_t size) {
     
     block = malloc(size);
     if (block == NULL) {
-        fprintf(stderr, "Failed to malloc %u bytes\n", size);
+        fprintf(stderr, "Failed to malloc %u bytes\n", (unsigned int) size);
         exit(EXIT_FAILURE);
     }
     return block;
